fix(npm): make local install paths prerelease-inclusive too
Address review of #34928: the npm installer's plain-req / npm:-import /
catalog resolution still used the prerelease-excluding matches_req, so a
prerelease workspace member referenced by a plain semver req resolved
locally under deno run/byonm but fell back to the registry under deno
install. Make those three sites prerelease-inclusive.

Consolidate the matching logic: move version_req_matches_including_pre to
deno_config (the lowest common crate), add NpmPackageConfig::
matches_req_including_pre, and re-export the helper from deno_resolver so
byonm/cache_deps imports are unchanged. workspace.rs now calls the shared
helper instead of an inline copy.

Add a deno-managed (nodeModulesDir auto) spec test covering a plain '*'
dependency on a prerelease workspace member.
