+++ libs/resolver/npm/mod.rs
@@ -10,7 +10,6 @@ use deno_error::JsError;
 use deno_maybe_sync::MaybeSend;
 use deno_maybe_sync::MaybeSync;
 use deno_maybe_sync::new_rc;
-use deno_semver::RangeSetOrTag;
 use deno_semver::Version;
 use deno_semver::npm::NpmPackageReqReference;
 use deno_semver::package::PackageReq;
@@ -690,62 +689,8 @@ pub(crate) fn join_package_name_to_path(
   path.into_owned()
 }
 
-/// Like `VersionReq::matches`, but also matches prerelease versions that fall
-/// within the requirement's range bounds.
-///
-/// This is used when matching a locally installed (e.g. pnpm/npm workspace)
-/// package against a requirement. npm's default semver rules exclude
-/// prereleases from ranges like `*` or `^1.0.0` to avoid silently selecting an
-/// unstable version from the registry, but a package that is already installed
-/// in `node_modules` was put there explicitly by the user, so a member with a
-/// prerelease version (e.g. `0.40.0-pre`) should still satisfy a bare
-/// `npm:<pkg>` (`*`) requirement instead of being rejected.
-///
-/// Unlike `VersionReq::matches`, a tag requirement never matches here (rather
-/// than panicking), since a local package has no dist-tags to compare against.
-pub fn version_req_matches_including_pre(
-  version_req: &deno_semver::VersionReq,
-  version: &Version,
-) -> bool {
-  match version_req.inner() {
-    RangeSetOrTag::RangeSet(set) => {
-      set.satisfies(version)
-        || (!version.pre.is_empty()
-          && set.0.iter().any(|range| range.intersects_version(version)))
-    }
-    RangeSetOrTag::Tag(_) => false,
-  }
-}
-
-#[cfg(test)]
-mod test {
-  use deno_semver::VersionReq;
-
-  use super::*;
-
-  #[test]
-  fn version_req_matches_including_pre_cases() {
-    fn matches(req: &str, version: &str) -> bool {
-      version_req_matches_including_pre(
-        &VersionReq::parse_from_npm(req).unwrap(),
-        &Version::parse_from_npm(version).unwrap(),
-      )
-    }
-
-    // a prerelease version satisfies a wildcard, unlike plain npm semver
-    assert!(matches("*", "0.40.0-pre"));
-    // ... and an explicit prerelease range that contains it
-    assert!(matches("^0.40.0-pre", "0.40.0-pre"));
-    // but not a range whose lower bound is above the prerelease
-    assert!(!matches("^0.40.0", "0.40.0-pre"));
-    // an exact version does not match a lower prerelease of itself
-    assert!(!matches("1.3.6", "1.3.6-beta"));
-    // a prerelease below the exclusive upper bound is included
-    assert!(matches("^1.0.0", "2.0.0-pre"));
-    // non-prerelease behaviour is unchanged
-    assert!(matches("^1.0.0", "1.5.0"));
-    assert!(!matches("^1.0.0", "2.0.0"));
-    // a tag never matches a local package (and does not panic)
-    assert!(!matches("latest", "1.0.0"));
-  }
-}
+// Matching a locally installed (e.g. pnpm/npm workspace) package against a
+// requirement is prerelease-inclusive. The canonical implementation lives in
+// `deno_config` so it can be shared with `NpmPackageConfig` matching; re-export
+// it here for the byonm and cache_deps callers.
+pub use deno_config::workspace::version_req_matches_including_pre;
