+++ tests/specs/npm/workspace_prerelease_member_local/__test__.jsonc
@@ -1,0 +1,16 @@
+{
+  "tempDir": true,
+  "steps": [
+    {
+      // a plain `*` dependency on a prerelease workspace member (not
+      // `workspace:*`) must resolve to the local member with deno-managed
+      // node_modules, instead of being fetched from the registry.
+      "args": "install",
+      "output": "[WILDCARD]"
+    },
+    {
+      "args": "run app/main.ts",
+      "output": "main.out"
+    }
+  ]
+}
