+++ tests/specs/npm/workspace_prerelease_member_local/app/main.ts
@@ -1,0 +1,3 @@
+import { hello } from "@denotest/prerelease-member";
+
+hello();
