+++ tests/specs/npm/workspace_prerelease_member_local/app/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@denotest/app",
+  "dependencies": {
+    "@denotest/prerelease-member": "*"
+  }
+}
