+++ tests/specs/npm/workspace_prerelease_member_local/deno.json
@@ -1,0 +1,4 @@
+{
+  "workspace": ["./member", "./app"],
+  "nodeModulesDir": "auto"
+}
