+++ tests/specs/npm/workspace_prerelease_member_local/main.out
@@ -1,0 +1,1 @@
+[WILDCARD]hello from local prerelease member
