+++ tests/specs/npm/workspace_prerelease_member_local/member/mod.js
@@ -1,0 +1,3 @@
+export function hello() {
+  console.log("hello from local prerelease member");
+}
