+++ tests/specs/npm/workspace_prerelease_member_local/member/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@denotest/prerelease-member",
+  "version": "0.40.0-pre",
+  "exports": "./mod.js"
+}
